backend/requirements.txt
fastapi
uvicorn
python-multipart
pydantic
google-generativeai>=0.3.0
python-dotenv
